// src/store/modules/app/types.ts
export enum AppMutationTypes {
  TOGGLE_SIDEBAR = "TOGGLE_SIDEBAR",
  CLOSE_SIDEBAR = "CLOSE_SIDEBAR",
  TOGGLE_DEVICE = "TOGGLE_DEVICE",
  SET_LANGUAGE = "SET_LANGUAGE",
  SET_SIZE = "SET_SIZE",
  SET_THEME = "SET_THEME",
}

export enum AppActionTypes {
  ACTION_TOGGLE_SIDEBAR = "ACTION_TOGGLE_SIDEBAR",
  ACTION_CLOSE_SIDEBAR = "ACTION_CLOSE_SIDEBAR",
  ACTION_TOGGLE_DEVICE = "ACTION_TOGGLE_DEVICE",
  ACTION_SET_LANGUAGE = "ACTION_SET_LANGUAGE",
  ACTION_SET_SIZE = "ACTION_SET_SIZE",
  ACTION_SET_THEME = "ACTION_SET_THEME",
}